requirements.txt
alembic==1.10.3
anyio==3.6.2
asgiref==3.6.0
asyncpg==0.27.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
dnspython==2.3.0
drf-yasg==1.21.5
drf-yasg2==1.19.4
email-validator==1.3.1
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.3.0
importlib-resources==5.12.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
orjson==3.8.9
packaging==23.1
pydantic==1.10.7
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.8
sqlparse==0.4.3
starlette==0.26.1
typing-extensions==4.5.0
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0
zipp==3.15.0
